requirements.txt
cython==0.29.13
snowflake-connector-python==2.5.1
python-dotenv==0.19.0
pyarrow==3.0.0
tqdm==4.62.3
tqdm
lxml